Replace outdated dependencies
The error macro crate could also be replaced with `thiserror` but I
prefer the simple macro to proc-macro dependencies. This removes some of
the types, i.e. no more type-per-variant, however these are mostly
internal and unusual in the ecosystem.

The network enumeration crate is implemented much better by the `netdev`
crate. In particular for IPv6 scopes it is insufficient to list the
interface addresses only or to assume one address per interface. The
internal logic is not updated in this commit.