Migrate away from hyper
The ecosystem has moved on to http/httpparse/headers which splits
responsibility much better. We only need the parser portion of HTTP/1.1.
Indeed the IO part of hyper is standing in our way. For instance the
network connections take a protocol whereas our connectors need to use
the right IP-based protocol (UDP or TCP) based on content. The trait is
thus misfit here. Additionally, the async/sync split affects us too and
this is a much more sans-IO core than previously.

This is incomplete, the parsing code will not yet verify the body of a
parsed request and response as it should. Interestingly, the NOTIFY
method is a request with a body.